Move outgoing networking to separate file
